Move middleware into better structure. Fix imports. Comments etc.
