+++ composer.json
@@ -7,6 +7,7 @@
     "require": {
         "php": ">=5.5.9",
         "illuminate/contracts": "^5.1|^5.2",
+        "illuminate/http": "^5.1|^5.2",
         "illuminate/support": "^5.1|^5.2"
     },
     "autoload": {
